Fix(Text Formatting): Remove indentation of first paragraph
The first paragraph of `\[...]section` has an indentation as the other paragraphs.
This is a non standard behaviour. Package `titlesec` tells about `\titlespacing`
command:
"The starred version kills the indentation of the paragraph following the title..."
So the fix adds the star to `\titlespacing` commands used in FEIstyle.cls file.